+++ core/src/main/java/com/ray3k/gdxparticleeditor/Settings.java
@@ -33,6 +33,7 @@ public class Settings {
     public static final int[] DEFAULT_REDO_SECONDARY_KEYBIND = {Keys.CONTROL_LEFT, Keys.SHIFT_LEFT, Keys.Z};
     public static final int[] DEFAULT_SAVE_PRIMARY_KEYBIND = {Keys.CONTROL_LEFT, Keys.S};
     public static final int[] DEFAULT_SAVE_AS_PRIMARY_KEYBIND = {Keys.CONTROL_LEFT, Keys.SHIFT_LEFT, Keys.S};
+    public static final int[] DEFAULT_EXPORT_PRIMARY_KEYBIND = {Keys.CONTROL_LEFT, Keys.E};
     public static final int[] DEFAULT_OPEN_PRIMARY_KEYBIND = {Keys.CONTROL_LEFT, Keys.O};
     public static final int[] DEFAULT_SWITCH_MODE_KEYBIND = {Keys.CONTROL_LEFT, Keys.TAB};
     public static FileHandle logFile;